Added ZXBASIC integration
Added ZXBASIC integration

+++ editor.py
@@ -42,12 +42,17 @@ class Notebook(ttk.Notebook):
 class Tab(ttk.Frame):
     def __init__(self, *args, FileDir):
         ttk.Frame.__init__(self, *args)
+        self.status_bar = None
         self.textbox = self.create_text_widget()
         self.file_dir = None
         self.file_name = os.path.basename(FileDir)
         self.status = md5(self.textbox.get(1.0, 'end').encode('utf-8'))
 
     def create_text_widget(self):
+        #Add the status bar
+        self.status_bar = tk.Label(self, text="", padx=5, relief="ridge", fg='white', bg="blue", anchor='w')
+        self.status_bar.pack(side='bottom', fill='x')
+
         # Horizontal Scroll Bar
         xscrollbar = tk.Scrollbar(self, orient='horizontal')
         xscrollbar.pack(side='bottom', fill='x')
@@ -83,7 +88,7 @@ class Editor:
         self.init_dir = os.path.join(os.path.expanduser('~'), 'Desktop')
         self.untitled_count = 1
 
-        # Create Notebook ( for tabs ).
+        # Create Notebook ( for tabs )
         self.nb = Notebook(self.master)
         self.nb.bind("<Button-2>", self.close_tab)
         self.nb.bind('<<NotebookTabChanged>>', self.tab_change)
@@ -92,40 +97,50 @@ class Editor:
         # Override the X button.
         self.master.protocol('WM_DELETE_WINDOW', self.exit)
 
-        # Create Menu Bar
+        # Menu Bar
         menubar = tk.Menu(self.master)
 
-        # Create File Menu
-        filemenu = tk.Menu(menubar, tearoff=0)
-        filemenu.add_command(label="New", command=self.new_file)
-        filemenu.add_command(label="Open", command=self.open_file)
-        filemenu.add_command(label="Save", command=self.save_file)
-        filemenu.add_command(label="Save As...", command=self.save_as)
-        filemenu.add_command(label="Close", command=self.close_tab)
-        filemenu.add_separator()
-        filemenu.add_command(label="Run", command=self.run_file)
-        filemenu.add_separator()
-        filemenu.add_command(label="Exit", command=self.exit)
+        # File Menu
+        file_menu = tk.Menu(menubar, tearoff=0)
+        file_menu.add_command(label="New", command=self.new_file)
+        file_menu.add_command(label="Open", command=self.open_file)
+        file_menu.add_command(label="Save", command=self.save_file)
+        file_menu.add_command(label="Save As...", command=self.save_as)
+        file_menu.add_command(label="Close", command=self.close_tab)
+        file_menu.add_separator()
+        file_menu.add_command(label="Exit", command=self.exit)
 
-        # Create Edit Menu
-        editmenu = tk.Menu(menubar, tearoff=0)
-        editmenu.add_command(label="Undo", command=self.undo)
-        editmenu.add_separator()
-        editmenu.add_command(label="Cut", command=self.cut)
-        editmenu.add_command(label="Copy", command=self.copy)
-        editmenu.add_command(label="Paste", command=self.paste)
-        editmenu.add_command(label="Delete", command=self.delete)
-        editmenu.add_command(label="Select All", command=self.select_all)
+        # Edit Menu
+        edit_menu = tk.Menu(menubar, tearoff=0)
+        edit_menu.add_command(label="Undo", command=self.undo)
+        edit_menu.add_separator()
+        edit_menu.add_command(label="Cut", command=self.cut)
+        edit_menu.add_command(label="Copy", command=self.copy)
+        edit_menu.add_command(label="Paste", command=self.paste)
+        edit_menu.add_command(label="Delete", command=self.delete)
+        edit_menu.add_command(label="Select All", command=self.select_all)
 
-        # Create Format Menu, with a check button for word wrap.
-        formatmenu = tk.Menu(menubar, tearoff=0)
+        # Format Menu
+        format_menu = tk.Menu(menubar, tearoff=0)
         self.word_wrap = tk.BooleanVar()
-        formatmenu.add_checkbutton(label="Word Wrap", onvalue=True, offvalue=False, variable=self.word_wrap, command=self.wrap)
+        format_menu.add_checkbutton(label="Word Wrap", onvalue=True, offvalue=False, variable=self.word_wrap, command=self.wrap)
+
+        # zmakebas Menu
+        zmake_menu = tk.Menu(menubar, tearoff=0)
+        zmake_menu.add_command(label="Run", command=self.run_zmakebas)
+
+        # zxbasic Menu
+        zxbasic_menu = tk.Menu(menubar, tearoff=0)
+        zxbasic_menu.add_command(label="Compile and Run", command=self.run_zxbasic)
+        zxbasic_menu.add_command(label="Show assembly", command=self.run_zxasm)
 
         # Attach to Menu Bar
-        menubar.add_cascade(label="File", menu=filemenu)
-        menubar.add_cascade(label="Edit", menu=editmenu)
-        menubar.add_cascade(label="Format", menu=formatmenu)
+        menubar.add_cascade(label="File", menu=file_menu)
+        menubar.add_cascade(label="Edit", menu=edit_menu)
+        menubar.add_cascade(label="Format", menu=format_menu)
+        menubar.add_cascade(label="zmakebas", menu=zmake_menu)
+        menubar.add_cascade(label="zxbasic", menu=zxbasic_menu)
+
         self.master.configure(menu=menubar)
 
         # Create right-click menu.
@@ -151,6 +166,10 @@ class Editor:
         self.master.bind_class('Text', '<Control-w>', self.close_tab)
         self.master.bind_class('Text', '<Button-3>', self.right_click)
 
+        #All below bindings are 1 position behind
+        self.master.bind("<Button 1>", self.update_status_bar)
+        self.master.bind("<Key>", self.update_status_bar)
+
         self.master.bind('<Configure>', self.configure_callback)
 
         # Create initial tab and 'Add' tab
@@ -161,6 +180,17 @@ class Editor:
         else:
             self.nb.add(Tab(self.master, FileDir='Untitled'), text='Untitled')
         self.nb.add(Tab(self.master, FileDir='f'), text=' + ')
+
+    def update_status_bar(self, event=None):
+        # schedule an update after the key press is handled
+        self.master.after(0, self._update_status_bar)
+
+    def _update_status_bar(self):
+        #Get cursor position
+        curr_tab = self.nb.current_tab()
+        cursor_pos =  curr_tab.textbox.index(tk.INSERT)
+        rc = cursor_pos.split('.')
+        curr_tab.status_bar.configure(text=f"Ln {rc[0]}, Col {rc[1]}")
 
     def configure_callback(self, event):
         Config.set('app.editor.geometry', self.master.geometry())
@@ -349,7 +379,8 @@ class Editor:
 
         # Exit if last tab is closed
         if self.nb.index("end") <= 1:
-            self.parent.editor = None
+            if self.parent:
+                self.parent.editor = None
             self.master.destroy()
 
         return True
@@ -359,7 +390,8 @@ class Editor:
         for i in range(self.nb.index('end')-1):
             if self.close_tab() is False:
                 break
-        self.parent.editor = None
+        if self.parent:
+            self.parent.editor = None
 
     def save_changes(self, tab):
         # Check if any changes have been made, returns False if user chooses to cancel rather than select to save or not.
@@ -391,22 +423,53 @@ class Editor:
         if self.nb.select() == self.nb.tabs()[-1]:
             self.new_file()
 
-    def run_file(self):
+    def run_zmakebas(self):
         curr_tab = self.nb.current_tab()
-        temp_bas_file = './tmp/tmp.bas'
-        temp_tap_file = './tmp/tmp.tap'
+        temp_bas_file = './tmp/tmpzmake.bas'
+        temp_tap_file = './tmp/tmpzmake.tap'
         file = open(temp_bas_file,"w")
         file.write(curr_tab.textbox.get(1.0, 'end'))
         file.close()
-        basic2tape.zmakebas(temp_bas_file, temp_tap_file)
-        self.parent.emulator.tape_load(temp_tap_file)
+        ret = basic2tape.zmakebas(temp_bas_file, temp_tap_file)
+        if ret and self.parent:
+            self.parent.emulator.tape_load(temp_tap_file)
+            os.remove(temp_bas_file)
+            os.remove(temp_tap_file)
+            self.parent.window.focus_force()
+
+    def run_zxbasic(self):
+        curr_tab = self.nb.current_tab()
+        temp_bas_file = './tmp/tmpzxbc.bas'
+        temp_tap_file = './tmp/tmpzxbc.tzx'
+        file = open(temp_bas_file,"w")
+        file.write(curr_tab.textbox.get(1.0, 'end'))
+        file.close()
+        ret, out_format, errors = basic2tape.zxbasic(temp_bas_file, temp_tap_file)
+        if ret:
+            self.open_text(errors)
+        elif self.parent:
+            self.parent.emulator.tape_load(temp_tap_file)
+            os.remove(temp_bas_file)
+            os.remove(temp_tap_file)
+            self.parent.window.focus_force()
+
+    def run_zxasm(self):
+        curr_tab = self.nb.current_tab()
+        temp_bas_file = './tmp/tmpzxbc.bas'
+        temp_asm_file = './tmp/tmpzxbc.asm'
+        file = open(temp_bas_file, "w", encoding="utf-8")
+        file.write(curr_tab.textbox.get(1.0, 'end'))
+        file.close()
+        ret, out_format, errors = basic2tape.zxasm(temp_bas_file, temp_asm_file)
+        if ret:
+            self.open_text(errors)
+        self.open_file(temp_asm_file)
         os.remove(temp_bas_file)
-        os.remove(temp_tap_file)
-        self.parent.window.focus_force()
+        os.remove(temp_asm_file)
 
 def main():
     root = tk.Tk()
-    app = Editor(root)
+    app = Editor(root, None)
     root.mainloop()
 
 if __name__ == '__main__':
